Added dummy rendering of initial motion

+++ PhysicsEngine.Domain/Space/Position.cs
@@ -2,18 +2,18 @@
 {
     public sealed class Position
     {
-        public Position(float x = 0, float y = 0, float z = 0)
+        public Position(double x = 0, double y = 0, double z = 0)
         {
             X = x;
             Y = y;
             Z = z;
         }
 
-        public float X { get; set; }
+        public double X { get; set; }
 
-        public float Y { get; set; }
+        public double Y { get; set; }
 
-        public float Z { get; set; }
+        public double Z { get; set; }
 
         public override string ToString()
         {
